Fix issue with openweatherapi

+++ EInkService/OpenWeatherMap/GetOneCallApiResult.cs
@@ -47,7 +47,7 @@ namespace EInkService.OpenWeatherMap
     public class Minutely
     {
         public DateTime dt { get; set; }
-        public int precipitation { get; set; }
+        public float precipitation { get; set; }
     }
 
     public class Hourly
